requirement.txt
h5py==3.6.0
scipy==1.7.3
numpy==1.21.5
pandas==1.3.5
sklearn==1.0
matplotlib==3.5.1
neurokit2==0.1.7
pytorch==1.10.1+cu102
torchvision==0.11.2+cu102